+++ src/intlit.fu.cpp
@@ -50,7 +50,7 @@ s_Intlit Intlit(fu::view<fu::byte> sign_prefix_value_suffix)
     const fu::byte c_suffix = (value_suffix.size() ? value_suffix[(value_suffix.size() - 1)] : (*(const fu::byte*)fu::NIL));
     const fu::byte suffix = ((c_suffix == fu::byte('u')) ? c_suffix : (*(const fu::byte*)fu::NIL));
     fu::view<fu::byte> value = fu::get_view(value_suffix, 0, (value_suffix.size() - ((suffix != fu::byte{}) ? 1 : int{})));
-    const uint64_t base = ((prefix == fu::byte('x')) ? uint64_t(16) : ((prefix == fu::byte('o')) ? uint64_t(8) : ((prefix == fu::byte('b')) ? uint64_t(2) : uint64_t(10))));
+    const uint64_t base = ((prefix == fu::byte('x')) ? uint64_t(unsigned(16)) : ((prefix == fu::byte('o')) ? uint64_t(unsigned(8)) : ((prefix == fu::byte('b')) ? uint64_t(unsigned(2)) : uint64_t(unsigned(10)))));
     fu_STR error {};
     uint64_t absval {};
     for (int i = 0; i < value.size(); i++)
@@ -77,7 +77,7 @@ s_Intlit Intlit(fu::view<fu::byte> sign_prefix_value_suffix)
     const int minsize_f = ((absval < 0x1000000ull) ? 32 : 64);
     fu_STR* _2;
     (*(_2 = &(error)) ? *_2 : *_2 = (sIgned ? ((minsize_i > 64) ? "Oversized signed int literal."_fu : fu_STR{}) : ((minsize_u > 64) ? "Oversized unsigned int literal."_fu : fu_STR{})));
-    return s_Intlit { fu::u8(base), fu::u8(minsize_i), fu::u8(minsize_u), fu::u8(minsize_f), bool(sIgned), bool(uNsigned), bool(negative), uint64_t(absval), fu_STR(error) };
+    return s_Intlit { fu::u8(base), fu::u8(unsigned(minsize_i)), fu::u8(unsigned(minsize_u)), fu::u8(unsigned(minsize_f)), bool(sIgned), bool(uNsigned), bool(negative), uint64_t(absval), fu_STR(error) };
 }
 
 #endif
